+++ apps/cruncher-server/src/engineV2/localstate/migrations/0001_flat_dormammu.sql
@@ -1,0 +1,4 @@
+ALTER TABLE `query_history` ADD `events_result_path` text;--> statement-breakpoint
+ALTER TABLE `query_history` ADD `table_result_path` text;--> statement-breakpoint
+ALTER TABLE `query_history` ADD `table_result_columns` text;--> statement-breakpoint
+ALTER TABLE `query_history` ADD `view_result_path` text;