+++ apps/cruncher/src/processes/frontend/components/ui/progress.tsx
@@ -1,5 +1,4 @@
 import { Progress as ChakraProgress } from "@chakra-ui/react";
-import { InfoTip } from "./toggle-tip";
 import * as React from "react";
 
 export const ProgressBar = React.forwardRef<
@@ -14,23 +13,4 @@ export const ProgressBar = React.forwardRef<
   );
 });
 
-export interface ProgressLabelProps extends ChakraProgress.LabelProps {
-  info?: React.ReactNode;
-}
-
-export const ProgressLabel = React.forwardRef<
-  HTMLDivElement,
-  ProgressLabelProps
->(function ProgressLabel(props, ref) {
-  const { children, info, ...rest } = props;
-  return (
-    // @ts-expect-error - lib component
-    <ChakraProgress.Label {...rest} ref={ref}>
-      {children}
-      {info && <InfoTip>{info}</InfoTip>}
-    </ChakraProgress.Label>
-  );
-});
-
 export const ProgressRoot = ChakraProgress.Root;
-export const ProgressValueText = ChakraProgress.ValueText;