+++ packages/adapters/grafana-loki-browser/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.3.7
+
+### Patch Changes
+
+- [#41](https://github.com/IamShobe/cruncher/pull/41) [`73a494e`](https://github.com/IamShobe/cruncher/commit/73a494e4f16d067dcc82100ac29fa2dcf2e77a4d) Thanks [@IamShobe](https://github.com/IamShobe)! - Engine revamped with disk-composed queries: the server process is now a standalone package (`cruncher-server`) with shared types in `@cruncher/server-shared`, enabling better separation of concerns and disk-based query composition.
+
+- Updated dependencies [[`73a494e`](https://github.com/IamShobe/cruncher/commit/73a494e4f16d067dcc82100ac29fa2dcf2e77a4d)]:
+  - @cruncher/qql@0.4.5
+  - @cruncher/adapter-loki@0.2.7
+  - @cruncher/adapter-utils@0.5.2
+
 ## 0.3.6
 
 ### Patch Changes